+++ Piraeus.BetterHistoryMod/Model/RunMeta.cs
@@ -30,7 +30,7 @@ public class RunMeta
     public int? Floor { get; set; }
 
     [JsonPropertyName("final_coins")]
-    public long FinalCoins { get; set; }
+    public double FinalCoins { get; set; }
 
     [JsonPropertyName("total_spins")]
     public int TotalSpins { get; set; }
